+++ WaveTablePlugin/Source/Synthesizer.h
@@ -16,53 +16,53 @@
 class WaveTableSound : public juce::SynthesiserSound
 {
 public:
-    WaveTableSound(std::vector<double>& table);
+    WaveTableSound(std::vector<float>& table);
     bool appliesToNote(int midiNoteNumber) override;
     bool appliesToChannel(int midiChannel) override;
-    std::vector<double>& getTable();
+    std::vector<float>& getTable();
     
 private:
-    std::vector<double>& m_table;
+    std::vector<float>& m_table;
 };
 
 class SynthVoice : public juce::SynthesiserVoice
 {
 public:
-    SynthVoice(std::vector<double>& table, std::vector<double>& table2, int tSize);
+    SynthVoice(std::vector<float>& table, std::vector<float>& table2, int tSize);
     
     bool canPlaySound(juce::SynthesiserSound*) override;
     void startNote(int midiNoteNumber, float velocity, juce::SynthesiserSound *sound, int currentPitchWheelPosition) override;
     void stopNote(float velocity, bool allowTailOff) override;
     
     void renderNextBlock(juce::AudioBuffer<float> &outputBuffer, int startSample, int numSamples) override;
-    double interpNextSamp() noexcept;
-    double interpNextSamp(std::vector<double>& table) noexcept;
+    float interpNextSamp() noexcept;
+    float interpNextSamp(std::vector<float>& table) noexcept;
     void updateAngle();
     
     //Unused Pure Virtual Functions
     void pitchWheelMoved(int newPitchWheelValue) override;
     void controllerMoved(int controllerNumber, int newControllerValue) override;
     
-    double interpolateValue(float val);
+    float interpolateValue(float val);
     
     void setParameters(std::atomic<float>* param);
     
     
 private:
-    const double tailDecayTime = 1.0;
+    const float tailDecayTime = 1.0;
     
-    double currentIndex;
-    double m_angle;
-    double m_angleDelta;
-    double m_level;
-    double m_freq;
-    double m_tail;
-    double m_tailDec = 0.99;
+    float currentIndex;
+    float m_angle;
+    float m_angleDelta;
+    float m_level;
+    float m_freq;
+    float m_tail;
+    float m_tailDec = 0.99;
     
     std::atomic<float>* interpParam = nullptr;
     
-    std::vector<double>& m_table;
-    std::vector<double>& m_table2;
+    std::vector<float>& m_table;
+    std::vector<float>& m_table2;
     unsigned int m_tableSize;
 };
